SRF support removed, cleaning up unneeded cruft.  Expect a new SRF support that will actually work as intended in all cases.
I made a few more small changes, and the base support is looking good.
Arrays now work.  The rest of the thing has been already working for a
while as well.

Also, add the "expected" files to the regression test suite, so that the
test report as "passed" as things start to look better.  The trigger file
contains one failure, which I introduced voluntarily until pg_raise works.