Convert the SPI functions to using a PHP resource instead of an array. This increases reliability as there's no way for a user to screw us by passing an invalid pointer.  Additionally, PHP takes care of freeing the resources automatically when it's no longer referenced.
There's still a memory leak in spi_fetch_row that needs to be fixed
before this code can process huge query results.

To support this, three new functions had to be added: spi_processed returns
the number of tuples returned by a query; spi_status returns the string
representation of the return value of the SPI_exec call; and spi_rewind
allows to restart the scanning of a result set.

More functions are still needed for a complete SPI stack, but they are quite
easy to write now that I've figured out how this thing works.