Fix error reporting so that errors generated in PHP can be passed to Postgres, and errors from SPI are passed to PHP so it can clean up after itself.
This commit means that this thing is finally starting to resemble something
usable.  I'm updating the expected files for the SPI and "raise" tests, as
both are producing exactly what I expect.  This means we are currently, for
the first time, passing all tests.

(There are some more tests I want to add to the suite before calling this the
1.1 release.  But we are getting very close.)