Add tests for the pseudotype and composite type functionalities.  Fixes #13.
Fix a very small issue in the validator that prevented it from running at all.
Also make sure all the PHP initialization code is inside a zend_try/catch
block.  Not sure this makes much sense at all because we set the error
handling callback only at the very end of this process, but ...